pyproject.toml
[project]
name = "yoyopod"
dynamic = ["version"]
description = "YoYoPod integrated VoIP and local-first mpv player for Raspberry Pi"
readme = "README.md"
requires-python = ">=3.12"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
authors = [
    { name = "YoYoPod Team" }
]
keywords = ["yoyopod", "raspberry-pi", "voip", "mpv", "audio-player", "local-first"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
urls = { Repository = "https://github.com/moustafattia/yoyopod-core", Issues = "https://github.com/moustafattia/yoyopod-core/issues" }

dependencies = [
    "cffi>=1.17.0",
    "displayhatmini>=0.0.2; platform_system == 'Linux'",
    "pygame>=2.5.0",
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "gpiod<2; platform_system == 'Linux'",
    "pyyaml>=6.0.3",
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "python-socketio>=5.11.0",
    "pynput>=1.7.0",
    "tinytag>=2.2.1",
    "pyserial>=3.5",
    "typer>=0.12.0",
    "paho-mqtt>=2.0.0",
    "msgpack>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.4.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML",
    "pyinstrument>=5.0.0",
    "pyperf>=2.8.0",
    "py-spy>=0.4.1",
]

[project.scripts]
yoyopod = "yoyopod_cli.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "yoyopod_cli/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["yoyopod", "yoyopod_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/README.md",
    "/pyproject.toml",
    "/yoyopod",
    "/yoyopod.py",
    "/yoyopod_cli",
]
exclude = [
    "/.agents",
    "/.claude",
    "/.worktrees",
    "/dist",
    "/logs",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "yoyopod_cli.pi.support.cloud_backend",
    "yoyopod_cli.pi.support.cloud_backend.*",
    "yoyopod_cli.pi.support.cloud_integration",
    "yoyopod_cli.pi.support.cloud_integration.*",
    "yoyopod_cli.pi.support.contacts_integration",
    "yoyopod_cli.pi.support.contacts_integration.*",
    "yoyopod_cli.pi.support.display",
    "yoyopod_cli.pi.support.display.*",
    "yoyopod_cli.pi.support.gpiod_compat",
    "yoyopod_cli.pi.support.lvgl_binding",
    "yoyopod_cli.pi.support.lvgl_binding.*",
    "yoyopod_cli.pi.support.music_backend",
    "yoyopod_cli.pi.support.music_backend.*",
    "yoyopod_cli.pi.support.music_integration",
    "yoyopod_cli.pi.support.music_integration.*",
    "yoyopod_cli.pi.support.power_backend",
    "yoyopod_cli.pi.support.power_backend.*",
    "yoyopod_cli.pi.support.power_integration",
    "yoyopod_cli.pi.support.power_integration.*",
    "yoyopod_cli.pi.support.voip_backend",
    "yoyopod_cli.pi.support.voip_backend.*",
]
ignore_errors = true

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/legacy_python_runtime"]

[tool.yoyopod_quality]
gate_format_paths = [
    "scripts/quality.py",
    "yoyopod_cli",
]
gate_lint_paths = [
    "scripts/quality.py",
    "yoyopod_cli",
]
gate_type_paths = [
    "scripts/quality.py",
    "yoyopod_cli",
]
audit_format_paths = ["."]
audit_lint_paths = ["."]
audit_type_paths = ["yoyopod"]
